requirements.txt
git+https://github.com/facebookresearch/segment-anything.git
torch>=2.0.0
torchvision~=0.16.2
segmentation-models-pytorch
Pillow~=9.5.0
tqdm
timm~=0.9.2
wandb~=0.15.12
numpy~=1.25.2
pandas
matplotlib~=3.7.3
seaborn
scikit-learn~=1.3.0
lightning~=2.0.9.post0
wget~=3.2
wheel~=0.38.4
setuptools~=65.5.1
packaging~=23.0
zipp~=3.15.0
future~=0.18.2
platformdirs~=3.5.1
datasets~=2.14.5
PyYAML~=6.0.1
transformers~=4.41.2
opencv-python
label-studio
label-studio-converter
prettytable
coco-eval
albumentations
accelerate